refactors generation for Mapped resource
A mapped resource is a resource for which the generator
creates classes that convert the resource data into a target pojo.
The target pojo may be a jpa entity, but not necessarily.
The target pojo ought to have a controller, and there ought to
be mappers (to map data to and from the resource and target).